move all utility functions to data.utils
