+++ blackdoc/tests/data/rst.py
@@ -1,4 +1,4 @@
-from . import from_dict
+from .utils import from_dict
 
 content = """\
 Long description of the function's assumptions and on how to call it.
